edit admin_menu function

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -13,26 +13,28 @@ class CreateUsersTable extends Migration
      */
     public function up()
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string('name');
-            $table->string('email')->unique();
-            $table->timestamp('email_verified_at')->nullable();
-            $table->string('password');
-            $table->rememberToken();
-            $table->timestamps();
-        });
+        if (env('DB_REFRESH')) {
+            Schema::create('users', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->string('name');
+                $table->string('email')->unique();
+                $table->timestamp('email_verified_at')->nullable();
+                $table->string('password');
+                $table->rememberToken();
+                $table->timestamps();
+            });
 
-        //忘記密碼寄驗證信
-        Schema::create('user_verification', function (Blueprint $table) {
-            $table->increments( 'id' );
-            $table->integer( 'user_id' )->common('admin_id,member_id','user_id');
-            $table->string( 'user_table' )->common('admins,members','users');
-            $table->string( 'verification', 50);
-            $table->dateTime( 'start_time' );
-            $table->dateTime( 'end_time' );
-            $table->tinyInteger( 'status' )->default(0);
-        });
+            //忘記密碼寄驗證信
+            Schema::create('user_verification', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('user_id')->comment('admin_id,member_id,user_id');
+                $table->string('user_table')->comment('admins,members,users');
+                $table->string('verification', 50);
+                $table->dateTime('start_time')->nullable()->comment('開始時間');
+                $table->dateTime('end_time')->nullable()->comment('結束時間');
+                $table->tinyInteger('status')->default(0);
+            });
+        }
     }
 
     /**
@@ -42,7 +44,9 @@ class CreateUsersTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('user_verification');
-        Schema::dropIfExists('users');
+        if (env('DB_REFRESH')) {
+            Schema::dropIfExists('user_verification');
+            Schema::dropIfExists('users');
+        }
     }
 }
