+++ database/migrations/2019_06_17_150331_create_log_logins_table.php
@@ -13,14 +13,30 @@ class CreateLogLoginsTable extends Migration
      */
     public function up()
     {
-        Schema::create('log_logins', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->common('登入者id')->nullable();
-            $table->string('user_type')->common('登入者類')->nullable();
-            $table->string('action')->common('操作紀錄')->nullable();
-            $table->string('ip')->common('IP位置')->nullable();
-            $table->timestamps();
-        });
+        if (env('DB_REFRESH')) {
+            //紀錄登入
+            Schema::create('log_logins', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('user_id')->comment('登入者id')->nullable();
+                $table->string('user_type')->comment('登入者類')->nullable();
+                $table->string('action')->comment('操作紀錄')->nullable();
+                $table->string('ip')->comment('IP位置')->nullable();
+                $table->timestamps();
+            });
+
+            //紀錄操作動作
+            Schema::create('log_actions', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('user_id')->comment('登入者id')->nullable();
+                $table->string('user_type')->comment('登入者類')->nullable();
+                $table->string('table_id')->comment('表id')->nullable();
+                $table->string('table_name')->comment('表名稱')->nullable();
+                $table->string('action')->comment('操作紀錄')->nullable();
+                $table->longText('value')->comment('值')->nullable();
+                $table->string('ip')->comment('IP位置')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -30,6 +46,9 @@ class CreateLogLoginsTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('log_logins');
+        if (env('DB_REFRESH')) {
+            Schema::dropIfExists('log_actions');
+            Schema::dropIfExists('log_logins');
+        }
     }
 }
